Fix regression that restore #3 #1 (#58)

+++ liberty/TableModel.cc
@@ -1026,13 +1026,13 @@ Table2::findValue(float axis_value1,
       return value(0, 0);
     else {
       size_t axis_index2 = axis2_->findAxisIndex(axis_value2);
-      float x2 = axis_value2;
-      float y00 = value(0, axis_index2);
-      float x2l = axis2_->axisValue(axis_index2);
-      float x2u = axis2_->axisValue(axis_index2 + 1);
-      float dx2 = (x2 - x2l) / (x2u - x2l);
-      float y01 = value(0, axis_index2 + 1);
-      float tbl_value
+      double x2 = axis_value2;
+      double y00 = value(0, axis_index2);
+      double x2l = axis2_->axisValue(axis_index2);
+      double x2u = axis2_->axisValue(axis_index2 + 1);
+      double dx2 = (x2 - x2l) / (x2u - x2l);
+      double y01 = value(0, axis_index2 + 1);
+      double tbl_value
 	= (1 - dx2) * y00
 	+      dx2  * y01;
       return tbl_value;
@@ -1040,13 +1040,13 @@ Table2::findValue(float axis_value1,
   }
   else if (size2 == 1) {
     size_t axis_index1 = axis1_->findAxisIndex(axis_value1);
-    float x1 = axis_value1;
-    float y00 = value(axis_index1, 0);
-    float x1l = axis1_->axisValue(axis_index1);
-    float x1u = axis1_->axisValue(axis_index1 + 1);
-    float dx1 = (x1 - x1l) / (x1u - x1l);
-    float y10 = value(axis_index1 + 1, 0);
-    float tbl_value
+    double x1 = axis_value1;
+    double y00 = value(axis_index1, 0);
+    double x1l = axis1_->axisValue(axis_index1);
+    double x1u = axis1_->axisValue(axis_index1 + 1);
+    double dx1 = (x1 - x1l) / (x1u - x1l);
+    double y10 = value(axis_index1 + 1, 0);
+    double tbl_value
       = (1 - dx1) * y00
       +      dx1  * y10;
     return tbl_value;
@@ -1054,19 +1054,19 @@ Table2::findValue(float axis_value1,
   else {
     size_t axis_index1 = axis1_->findAxisIndex(axis_value1);
     size_t axis_index2 = axis2_->findAxisIndex(axis_value2);
-    float x1 = axis_value1;
-    float x2 = axis_value2;
-    float y00 = value(axis_index1, axis_index2);
-    float x1l = axis1_->axisValue(axis_index1);
-    float x1u = axis1_->axisValue(axis_index1 + 1);
-    float dx1 = (x1 - x1l) / (x1u - x1l);
-    float y10 = value(axis_index1 + 1, axis_index2);
-    float y11 = value(axis_index1 + 1, axis_index2 + 1);
-    float x2l = axis2_->axisValue(axis_index2);
-    float x2u = axis2_->axisValue(axis_index2 + 1);
-    float dx2 = (x2 - x2l) / (x2u - x2l);
-    float y01 = value(axis_index1, axis_index2 + 1);
-    float tbl_value
+    double x1 = axis_value1;
+    double x2 = axis_value2;
+    double y00 = value(axis_index1, axis_index2);
+    double x1l = axis1_->axisValue(axis_index1);
+    double x1u = axis1_->axisValue(axis_index1 + 1);
+    double dx1 = (x1 - x1l) / (x1u - x1l);
+    double y10 = value(axis_index1 + 1, axis_index2);
+    double y11 = value(axis_index1 + 1, axis_index2 + 1);
+    double x2l = axis2_->axisValue(axis_index2);
+    double x2u = axis2_->axisValue(axis_index2 + 1);
+    double dx2 = (x2 - x2l) / (x2u - x2l);
+    double y01 = value(axis_index1, axis_index2 + 1);
+    double tbl_value
       = (1 - dx1) * (1 - dx2) * y00
       +      dx1  * (1 - dx2) * y10
       +      dx1  *      dx2  * y11
